+++ framework/src/main/java/org/checkerframework/framework/type/MostlyNoElementQualifierHierarchy.java
@@ -14,8 +14,8 @@ import org.checkerframework.framework.util.QualifierKindHierarchy;
  * A {@link org.checkerframework.framework.type.QualifierHierarchy} where qualifiers may be
  * represented by annotations with elements, but most of the qualifiers do not have elements. In
  * contrast to {@link org.checkerframework.framework.type.ElementQualifierHierarchy}, this class
- * partially implements {@link #isSubtype(AnnotationMirror, AnnotationMirror)}, {@link
- * #leastUpperBound(AnnotationMirror, AnnotationMirror)}, and {@link
+ * partially implements {@link #isSubtype(AnnotationMirror, TypeMirror, AnnotationMirror,
+ * TypeMirror)}, {@link #leastUpperBound(AnnotationMirror, AnnotationMirror)}, and {@link
  * #greatestLowerBound(AnnotationMirror, AnnotationMirror)} and calls *WithElements when the result
  * cannot be computed from the meta-annotations {@link
  * org.checkerframework.framework.qual.SubtypeOf}.
@@ -53,7 +53,7 @@ public abstract class MostlyNoElementQualifierHierarchy extends ElementQualifier
   }
 
   @Override
-  public final boolean isSubtype(
+  public final boolean isSubtypeImpl(
       AnnotationMirror subAnno,
       TypeMirror subType,
       AnnotationMirror superAnno,